+++ src/app/modules/index.ts
@@ -4,5 +4,6 @@ export * from './gpxload/gpxload.module'
 export * from './alarms/alarms.module'
 export * from './skresources/resources.module'
 export * from './map/map.module'
+export * from './utils/utils.module'
 
 export * from './skstream/skstream.module'