+++ src/app/modules/utils/utils.module.ts
@@ -1,0 +1,15 @@
+/*****************************
+ * Utilities Module
+ *****************************/
+import { NgModule } from '@angular/core';
+
+import  { CoordsPipe } from './coords.pipe'
+
+@NgModule({
+    imports: [],    
+    declarations: [ CoordsPipe ],
+    exports: [ CoordsPipe ],
+    entryComponents: [ ], 
+    providers: []  
+})
+export class UtilitiesModule {}